backend/requirements.txt
Flask
Flask-Cors
openai-whisper
torch
torchvision
torchaudio
transformers
diffusers
opencv-python
numpy
pandas
streamlit
Pillow
scikit-learn
tqdm